feat: Configured and inserted TunesProvider in App.js
