+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "files.insertFinalNewline": false,
+  "files.trimFinalNewlines": true
+}
