+++ src/index.js
@@ -2,12 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { TunesProvider } from './utils/provider';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
+  <TunesProvider >
     <App />
-  </React.StrictMode>
+  </TunesProvider>
 );
 
 // If you want to start measuring performance in your app, pass a function
