+++ src/pages/Home.jsx
@@ -9,4 +9,3 @@ export default function Home() {
     </div>
   )
 }
-
