--- src/utils/context/Index.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const tunesContext = createContext({});
-
-export default tunesContext;
