+++ src/utils/context/index.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const TunesContext = createContext({});
+
+export default TunesContext;
