+++ src/utils/provider/index.js
@@ -1,0 +1,23 @@
+import React, { useContext, useState } from 'react';
+import PropTypes from 'prop-types';
+import TunesContext from '../context';
+
+export const TunesProvider = ({ children }) => {
+  const [data, setData] = useState({
+    data: {
+      users: [],
+    } 
+  });
+
+  return (
+    <TunesContext.Provider value={ { data, setData } }>
+      { children }
+    </TunesContext.Provider>
+  );
+};
+
+export const useUpdateContext = () => useContext(TunesContext);
+
+TunesProvider.propTypes = {
+  children: PropTypes.element,
+}.isRequired;
